+++ packages/ui/src/lib/dialog/dialog.component.spec.ts
@@ -1,7 +1,7 @@
 import { Component, signal } from '@angular/core';
 import { type ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { provideEagamiUi } from '../i18n/i18n.provider';
+import { provideEagamiUi } from '../eagami-ui.provider';
 import { EagamiI18nService } from '../i18n/i18n.service';
 import { frFR } from '../i18n/messages';
 import { DialogComponent, type DialogWidth } from './dialog.component';